feat(castor): rename api-server to just server
