+++ castor/service/README.md
@@ -6,20 +6,15 @@ __Running Castor service locally for development__
 
 ```bash
 docker-compose -f docker/docker-compose-local.yaml up -d
-sbt api-server/run
+sbt server/run
 ```
 
 ---
 
 ## Known limitations
 
-### OpenAPI codegen
+### Castor `worker` process embedded inside `server`
 
-The usage of `oneOf` / `anyOf` in OpenAPI specification doesn't map nicely to scala code using default template.
-This has to be eventually addressed by customizing `mustache` template.
-
-### Castor `worker` process embedded inside `api-server`
-
-Castor worker process which is responsible for async comminucation with other components is temporarily embedded inside `api-server`.
+Castor worker process which is responsible for async comminucation with other components is temporarily embedded inside `server`.
 This impose some restrictions on scalability as it requires some coordination of between `worker` instances.
 It should be addressed going forward.
