--- castor/service/api-server/src/main/scala/io/iohk/atala/castor/apiserver/Main.scala
@@ -1,7 +1,0 @@
-package io.iohk.atala.castor.apiserver
-
-import zio.*
-
-object Main extends ZIOAppDefault {
-  override def run = Modules.app
-}
