--- castor/service/api-server/src/main/scala/io/iohk/atala/castor/apiserver/http/OASModelPatches.scala
@@ -1,9 +1,0 @@
-package io.iohk.atala.castor.apiserver.http
-
-object OASModelPatches {
-
-  type DidType = String
-  type DidOperationType = String
-  type DidOperationStatus = String
-
-}
