+++ castor/service/server/openapi/generator-config/scala-akka-http-server/stringDirectives.mustache
