+++ castor/service/server/src/main/scala/io/iohk/atala/castor/server/Main.scala
@@ -1,0 +1,7 @@
+package io.iohk.atala.castor.server
+
+import zio.*
+
+object Main extends ZIOAppDefault {
+  override def run = Modules.app
+}
