+++ castor/service/server/src/main/scala/io/iohk/atala/castor/server/http/OASModelPatches.scala
@@ -1,0 +1,9 @@
+package io.iohk.atala.castor.server.http
+
+object OASModelPatches {
+
+  type DidType = String
+  type DidOperationType = String
+  type DidOperationStatus = String
+
+}
